requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.18
    # via
    #   httpx-aiohttp
    #   litellm
aiosignal==1.3.2
    # via aiohttp
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
attrs==25.3.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
cachetools==5.5.2
    # via
    #   google-auth
    #   streamlit
certifi==2025.4.26
    # via
    #   httpcore
    #   httpx
    #   requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.2
    # via requests
click==8.1.8
    # via
    #   litellm
    #   streamlit
cryptography==43.0.3
    # via
    #   pydrive2
    #   pyopenssl
distro==1.9.0
    # via openai
filelock==3.18.0
    # via huggingface-hub
frozenlist==1.6.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.5.1
    # via huggingface-hub
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
google-api-core==2.24.2
    # via google-api-python-client
google-api-python-client==2.170.0
    # via
    #   managementsim (pyproject.toml)
    #   pydrive2
google-auth==2.40.2
    # via
    #   google-api-core
    #   google-api-python-client
    #   google-auth-httplib2
    #   google-auth-oauthlib
google-auth-httplib2==0.2.0
    # via google-api-python-client
google-auth-oauthlib==1.2.2
    # via managementsim (pyproject.toml)
googleapis-common-protos==1.70.0
    # via google-api-core
h11==0.16.0
    # via httpcore
hf-xet==1.1.2
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httplib2==0.22.0
    # via
    #   google-api-python-client
    #   google-auth-httplib2
    #   oauth2client
httpx==0.28.1
    # via
    #   httpx-aiohttp
    #   litellm
    #   openai
httpx-aiohttp==0.1.4
    # via litellm
huggingface-hub==0.32.0
    # via tokenizers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.7.0
    # via litellm
jinja2==3.1.6
    # via
    #   altair
    #   litellm
    #   pydeck
jiter==0.10.0
    # via openai
jsonschema==4.23.0
    # via
    #   altair
    #   litellm
jsonschema-specifications==2025.4.1
    # via jsonschema
litellm==1.71.0
    # via managementsim (pyproject.toml)
markupsafe==3.0.2
    # via jinja2
multidict==6.4.4
    # via
    #   aiohttp
    #   yarl
narwhals==1.37.1
    # via altair
numpy==2.2.5
    # via
    #   pandas
    #   pydeck
    #   streamlit
oauth2client==4.1.3
    # via pydrive2
oauthlib==3.2.2
    # via requests-oauthlib
openai==1.82.0
    # via litellm
packaging==24.2
    # via
    #   altair
    #   huggingface-hub
    #   streamlit
pandas==2.2.3
    # via
    #   managementsim (pyproject.toml)
    #   streamlit
pillow==11.2.1
    # via streamlit
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
proto-plus==1.26.1
    # via google-api-core
protobuf==6.30.2
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
    #   streamlit
pyarrow==20.0.0
    # via streamlit
pyasn1==0.6.1
    # via
    #   oauth2client
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via
    #   google-auth
    #   oauth2client
pycparser==2.22
    # via cffi
pydantic==2.11.5
    # via
    #   litellm
    #   openai
pydantic-core==2.33.2
    # via pydantic
pydeck==0.9.1
    # via streamlit
pydrive2==1.21.3
    # via managementsim (pyproject.toml)
pyopenssl==24.2.1
    # via pydrive2
pyparsing==3.2.3
    # via httplib2
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.0
    # via litellm
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   pydrive2
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via tiktoken
requests==2.32.3
    # via
    #   google-api-core
    #   huggingface-hub
    #   requests-oauthlib
    #   streamlit
    #   tiktoken
requests-oauthlib==2.0.0
    # via google-auth-oauthlib
rpds-py==0.24.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via
    #   google-auth
    #   oauth2client
six==1.17.0
    # via
    #   oauth2client
    #   python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via
    #   anyio
    #   openai
streamlit==1.45.0
    # via managementsim (pyproject.toml)
tenacity==9.1.2
    # via streamlit
tiktoken==0.9.0
    # via litellm
tokenizers==0.21.1
    # via litellm
toml==0.10.2
    # via streamlit
tornado==6.5.1
    # via
    #   managementsim (pyproject.toml)
    #   streamlit
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   openai
typing-extensions==4.13.2
    # via
    #   altair
    #   anyio
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   referencing
    #   streamlit
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
uritemplate==4.1.1
    # via google-api-python-client
urllib3==2.4.0
    # via requests
watchdog==6.0.0
    # via managementsim (pyproject.toml)
yarl==1.20.0
    # via aiohttp
zipp==3.21.0
    # via importlib-metadata
